+++ CHANGELOG.md
@@ -85,7 +85,7 @@ Make sure to upgrade to v0.25.0 so that your node will continue to connect once 
 - `/api/v1/balance`, `/api/v1/transactions`, `/api/v1/outputs` and `/api/v1/blocks` accept the `POST` method so that large request bodies can be sent to the server, which would not fit in a `GET` query string
 - `/api/v1/explorer/address` is deprecated in favor of `/api/v1/transactions?verbose=1`
 - `/api/v1/network/connection*` connection object's field `"introduced"` replaced with field `"state"` which may have the values `"pending"`, `"connected"` or `"introduced"`
-- `/api/v1/network/connection*` field `"is_default_peer"` added to connection object to indicate if the peer is in the hardcoded list of default peers
+- `/api/v1/network/connection*` field `"is_trusted_peer"` added to connection object to indicate if the peer is in the hardcoded list of default peers
 - `/api/v1/network/connection*` field `"connected_at"` added to connection object
 - `/api/v1/network/connections` now includes incoming connections. Filters are added to query connections by state and direction